// frontend/plugins/welcome/src/api/models/EntProduct.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SUT SA Example API Playlist Vidoe
 * This is a sample server for SUT SE 2563
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    EntProductEdges,
    EntProductEdgesFromJSON,
    EntProductEdgesFromJSONTyped,
    EntProductEdgesToJSON,
} from './';

/**
 * 
 * @export
 * @interface EntProduct
 */
export interface EntProduct {
    /**
     * 
     * @type {EntProductEdges}
     * @memberof EntProduct
     */
    edges?: EntProductEdges;
    /**
     * ID of the ent.
     * @type {number}
     * @memberof EntProduct
     */
    id?: number;
    /**
     * ProductName holds the value of the "product_name" field.
     * @type {string}
     * @memberof EntProduct
     */
    productName?: string;
    /**
     * ProductPaymentOfYear holds the value of the "product_payment_of_year" field.
     * @type {number}
     * @memberof EntProduct
     */
    productPaymentOfYear?: number;
    /**
     * ProductPrice holds the value of the "product_price" field.
     * @type {number}
     * @memberof EntProduct
     */
    productPrice?: number;
    /**
     * ProductTime holds the value of the "product_time" field.
     * @type {number}
     * @memberof EntProduct
     */
    productTime?: number;
}

export function EntProductFromJSON(json: any): EntProduct {
    return EntProductFromJSONTyped(json, false);
}

export function EntProductFromJSONTyped(json: any, ignoreDiscriminator: boolean): EntProduct {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'edges': !exists(json, 'edges') ? undefined : EntProductEdgesFromJSON(json['edges']),
        'id': !exists(json, 'id') ? undefined : json['id'],
        'productName': !exists(json, 'product_name') ? undefined : json['product_name'],
        'productPaymentOfYear': !exists(json, 'product_payment_of_year') ? undefined : json['product_payment_of_year'],
        'productPrice': !exists(json, 'product_price') ? undefined : json['product_price'],
        'productTime': !exists(json, 'product_time') ? undefined : json['product_time'],
    };
}

export function EntProductToJSON(value?: EntProduct | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'edges': EntProductEdgesToJSON(value.edges),
        'id': value.id,
        'product_name': value.productName,
        'product_payment_of_year': value.productPaymentOfYear,
        'product_price': value.productPrice,
        'product_time': value.productTime,
    };
}