// frontend/plugins/welcome/src/api/models/EntOfficer.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SUT SA Example API Playlist Vidoe
 * This is a sample server for SUT SE 2563
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    EntOfficerEdges,
    EntOfficerEdgesFromJSON,
    EntOfficerEdgesFromJSONTyped,
    EntOfficerEdgesToJSON,
} from './';

/**
 * 
 * @export
 * @interface EntOfficer
 */
export interface EntOfficer {
    /**
     * 
     * @type {EntOfficerEdges}
     * @memberof EntOfficer
     */
    edges?: EntOfficerEdges;
    /**
     * ID of the ent.
     * @type {number}
     * @memberof EntOfficer
     */
    id?: number;
    /**
     * OfficerEmail holds the value of the "officer_email" field.
     * @type {string}
     * @memberof EntOfficer
     */
    officerEmail?: string;
    /**
     * OfficerName holds the value of the "officer_name" field.
     * @type {string}
     * @memberof EntOfficer
     */
    officerName?: string;
    /**
     * OfficerPassword holds the value of the "officer_password" field.
     * @type {string}
     * @memberof EntOfficer
     */
    officerPassword?: string;
}

export function EntOfficerFromJSON(json: any): EntOfficer {
    return EntOfficerFromJSONTyped(json, false);
}

export function EntOfficerFromJSONTyped(json: any, ignoreDiscriminator: boolean): EntOfficer {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'edges': !exists(json, 'edges') ? undefined : EntOfficerEdgesFromJSON(json['edges']),
        'id': !exists(json, 'id') ? undefined : json['id'],
        'officerEmail': !exists(json, 'officer_email') ? undefined : json['officer_email'],
        'officerName': !exists(json, 'officer_name') ? undefined : json['officer_name'],
        'officerPassword': !exists(json, 'officer_password') ? undefined : json['officer_password'],
    };
}

export function EntOfficerToJSON(value?: EntOfficer | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'edges': EntOfficerEdgesToJSON(value.edges),
        'id': value.id,
        'officer_email': value.officerEmail,
        'officer_name': value.officerName,
        'officer_password': value.officerPassword,
    };
}